requirements.txt
PySide6>=6.7,<7
numpy>=1.26,<2.0
faster-whisper==1.2.1
ctranslate2==4.4.0
sounddevice>=0.4.7
process-audio-capture>=1.0.0
soundfile>=0.12.1
python-docx>=1.1
psutil>=5.9
requests>=2.32
setuptools==75.8.0
nvidia-cublas-cu12>=12.4.5.8; platform_system == "Windows"
nvidia-cudnn-cu12==8.9.7.29; platform_system == "Windows"
nvidia-cuda-nvrtc-cu12>=12.4.127; platform_system == "Windows"
